Fix runtime errors: remove failing Prisma field comparison in inventory and fix oxygenLevel type mismatch in patients

+++ actions/patients.ts
@@ -135,7 +135,7 @@ export async function addPatientVitals(patientId: string, formData: FormData) {
                 bloodPressure,
                 pulse: parseInt(pulse),
                 temperature: parseFloat(temperature),
-                oxygenLevel: oxygenLevel ? parseFloat(oxygenLevel) : null,
+                oxygenLevel: oxygenLevel ? parseInt(oxygenLevel) : null,
             }
         });
